+++ packages/node/src/ethereum/api.ethereum.test.ts
@@ -6,6 +6,7 @@ import { EventEmitter2 } from '@nestjs/event-emitter';
 import {
   EthereumDatasourceKind,
   EthereumHandlerKind,
+  EthereumLogFilter,
   SubqlRuntimeDatasource,
 } from '@subql/types-ethereum';
 import { EthereumApi } from './api.ethereum';
@@ -38,10 +39,22 @@ describe('Api.ethereum', () => {
   let ethApi: EthereumApi;
   const eventEmitter = new EventEmitter2();
   let blockData: EthereumBlockWrapped;
-  beforeAll(async () => {
+  beforeEach(async () => {
     ethApi = new EthereumApi(HTTP_ENDPOINT, eventEmitter);
     await ethApi.init();
     blockData = await ethApi.fetchBlock(16258633, true);
+  });
+
+  it('Should format transaction in logs, and the transaction gas should be bigInt type', () => {
+    expect(typeof blockData.logs[0].transaction.gas).toBe('bigint');
+    expect(typeof blockData.logs[0].transaction.blockNumber).toBe('number');
+    expect(typeof blockData.logs[0].transaction.gasPrice).toBe('bigint');
+    expect(typeof blockData.logs[0].transaction.maxPriorityFeePerGas).toBe(
+      'bigint',
+    );
+    expect(typeof blockData.logs[0].transaction.transactionIndex).toBe(
+      'bigint',
+    );
   });
 
   it('Decode nested logs in transactions', async () => {
@@ -87,6 +100,47 @@ describe('Api.ethereum', () => {
     );
     expect(result.length).toBe(1);
   });
+
+  it('!null filter support for logs, expect to filter out', async () => {
+    const beamEndpoint = 'https://rpc.api.moonbeam.network';
+    ethApi = new EthereumApi(beamEndpoint, eventEmitter);
+    await ethApi.init();
+    const filter_1: EthereumLogFilter = {
+      topics: [
+        '0xddf252ad1be2c89b69c2b068fc378daa952ba7f163c4a11628f55a4df523b3ef',
+        undefined,
+        undefined,
+        '!null',
+      ],
+    };
+
+    const filter_2: EthereumLogFilter = {
+      topics: [
+        '0xddf252ad1be2c89b69c2b068fc378daa952ba7f163c4a11628f55a4df523b3ef',
+      ],
+    };
+
+    blockData = await ethApi.fetchBlock(4015990, true);
+    const transaction = blockData.transactions.find(
+      (tx) =>
+        tx.hash ===
+        '0xeb2e443f2d4e784193fa13bbbae2b85e6ee459e7b7b53f8ca098ffae9b25b059',
+    );
+    const erc20Transfers = transaction.logs.filter((log) => {
+      if (EthereumBlockWrapped.filterLogsProcessor(log, filter_2)) {
+        return log;
+      }
+    });
+    const erc721Transfers = transaction.logs.filter((log) => {
+      if (EthereumBlockWrapped.filterLogsProcessor(log, filter_1)) {
+        return log;
+      }
+    });
+
+    expect(erc20Transfers.length).toBe(7);
+    expect(erc721Transfers.length).toBe(2);
+  });
+
   it('Null filter support, for undefined transaction.to', async () => {
     const beamEndpoint = 'https://rpc.api.moonbeam.network';
     ethApi = new EthereumApi(beamEndpoint, eventEmitter);
@@ -165,5 +219,31 @@ describe('Api.ethereum', () => {
       }
     });
     expect(result.length).toBe(2);
+  });
+
+  it('Resolves the correct tags for finalization', async () => {
+    // Ethereum
+    expect((ethApi as any).supportsFinalization).toBeTruthy();
+
+    // Moonbeam
+    ethApi = new EthereumApi('https://rpc.api.moonbeam.network', eventEmitter);
+    await ethApi.init();
+
+    expect((ethApi as any).supportsFinalization).toBeTruthy();
+
+    // BSC
+    ethApi = new EthereumApi('https://bsc-dataseed.binance.org', eventEmitter);
+    await ethApi.init();
+
+    expect((ethApi as any).supportsFinalized).toBeFalsy();
+
+    // Polygon
+    ethApi = new EthereumApi(
+      'https://polygon.api.onfinality.io/public',
+      eventEmitter,
+    );
+    await ethApi.init();
+
+    expect((ethApi as any).supportsFinalized).toBeFalsy();
   });
 });