+++ packages/node/src/ethereum/ethers/json-rpc-batch-provider.ts
@@ -34,6 +34,8 @@ export class JsonRpcBatchProvider extends JsonRpcProvider {
     reject: (error: Error) => void;
   }>;
 
+  _chainIdCache: string | null = null;
+
   constructor(url: string | ConnectionInfo, network?: Networkish) {
     super(url, network);
   }
@@ -43,6 +45,10 @@ export class JsonRpcBatchProvider extends JsonRpcProvider {
   }
 
   send(method: string, params: Array<any>): Promise<any> {
+    if (method === 'eth_chainId' && this._chainIdCache !== null) {
+      return Promise.resolve(this._chainIdCache);
+    }
+
     const request = {
       method: method,
       params: params,
@@ -129,6 +135,27 @@ export class JsonRpcBatchProvider extends JsonRpcProvider {
           return;
         }
 
+        // BSC returns a 200 response with this when being rate limited, we need a special case for it
+        /*
+           [
+              {
+                jsonrpc: '2.0',
+                id: null,
+                error: {
+                  code: -32005,
+                  message: 'method eth_getLogs in batch triggered rate limit'
+                }
+              }
+            ]
+        */
+        if (result.length === 1 && result[0].id === null) {
+          batch.forEach((inflightRequest) => {
+            inflightRequest.reject(new Error(result[0].error?.message));
+          });
+          this.adjustBatchSize(false);
+          return;
+        }
+
         const resultMap = result.reduce((resultMap, payload) => {
           resultMap[payload.id] = payload;
           return resultMap;
@@ -138,7 +165,13 @@ export class JsonRpcBatchProvider extends JsonRpcProvider {
         // on whether it was a success or error
         batch.forEach((inflightRequest) => {
           const payload = resultMap[inflightRequest.request.id];
-          if (payload.error) {
+          if (!payload) {
+            inflightRequest.reject(
+              new Error(
+                `Missing payload in response for request ${inflightRequest.request.id}`,
+              ),
+            );
+          } else if (payload.error) {
             const error = new Error(payload.error.message);
             (<any>error).code = payload.error.code;
             (<any>error).data = payload.error.data;
@@ -155,6 +188,9 @@ export class JsonRpcBatchProvider extends JsonRpcProvider {
             }
             inflightRequest.reject(error);
           } else {
+            if (inflightRequest.request.method === 'eth_chainId') {
+              this._chainIdCache = payload.result;
+            }
             inflightRequest.resolve(payload.result);
           }
         });