+++ packages/node/CHANGELOG.md
@@ -4,8 +4,35 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-
 ## [Unreleased]
+
+### Add 
+- Added `!null` filter for logs (#135)
+
+### Updated
+- Updated retry logic for eth requests (#134)
+
+### Fixed
+- Fixed missing mmrQueryService in indexer module
+
+## [2.9.2] - 2023-07-12
+### Fixed
+- Sync with @subql/node-core@3.1.2 (#130)
+  - Various improvements for POI feature
+  - Handle RPC error for oversize block responses
+- Missing format transaction in log (#128)
+
+## [2.9.1] - 2023-07-07
+### Fixed
+- Sync with @subql/node-core@3.1.1, fixed Poi table missing mmr issue
+- Finalization check with BSC and improved rate limit handling (#126)
+
+## [2.9.0] - 2023-07-05
+### Fixed
+- Limit the number of calls to eth_chainId (#123)
+
+### Added
+- `store-cache-upper-limit` flag to limit the max size of the store cache. (#124)
 
 ## [2.8.0] - 2023-06-27
 ### Changed
@@ -156,7 +183,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Retry request when encouraging timeout/rate limit behaviours (#9)
 
 ## [0.1.0] - 2022-10-31
-[Unreleased]: https://github.com/subquery/subql-ethereum/compare/node-ethereum/2.8.0...HEAD
+[Unreleased]: https://github.com/subquery/subql-ethereum/compare/node-ethereum/2.9.2...HEAD
+[2.9.2]: https://github.com/subquery/subql-ethereum/compare/node-ethereum/2.9.1...node-ethereum/2.9.2
+[2.9.1]: https://github.com/subquery/subql-ethereum/compare/node-ethereum/2.9.0...node-ethereum/2.9.1
+[2.9.0]: https://github.com/subquery/subql-ethereum/compare/node-ethereum/2.8.0...node-ethereum/2.9.0
 [2.8.0]: https://github.com/subquery/subql-ethereum/compare/node-ethereum/2.6.1...node-ethereum/2.8.0
 [2.6.1]: https://github.com/subquery/subql-ethereum/compare/node/2.6.0...node-ethereum/2.6.1
 [2.6.0]: https://github.com/subquery/subql-ethereum/compare/node-ethereum/2.5.3...node-ethereum/2.6.0